+++ python/ingestion/gcs_to_bq_util.py
@@ -9,7 +9,6 @@ from typing import List
 from ingestion.constants import BQ_STRING, BQ_FLOAT
 
 DATA_DIR = os.path.join(os.sep, 'app', 'data')
-ahr_api_key = os.getenv("AHR_API_KEY")
 
 
 def __convert_frame_to_json(frame):
@@ -408,73 +407,3 @@ def get_bq_column_types(df, float_cols: List[str]):
         column_types[col] = BQ_FLOAT
 
     return column_types
-
-
-def fetch_ahr_data_from_graphql():
-    """
-    Fetches data from AmericasHealthRankings GraphQL API.
-
-    Parameters:
-        query (str): The GraphQL query to be executed.
-        variables (dict): Variables to be passed with the query.
-    Returns:
-        a dictionary containing the data retrieved from the API.
-    """
-
-    graphql_url = 'https://api.americashealthrankings.org/graphql'
-    headers = {'Content-Type': 'application/json', 'x-api-key': ahr_api_key}
-    all_responses = []
-
-    query = """
-    query Data_A($where: MeasureFilterInput_A) {
-        measures_A(where: $where) {
-            data {
-                endDate
-                state
-                value
-                measure {
-                    name
-                }
-            }
-        }
-    }
-    """
-
-    descriptions_list = [
-        'asthma',
-        'cost',
-        'angina or coronary heart disease, a heart attack or myocardial infarction, or a stroke',
-        'kidney disease',
-        'chronic obstructive pulmonary disease',
-        'depression',
-        'diabetes',
-        'binge drinking',
-        'mental health',
-        'Discharges following hospitalization',
-        'prescription drugs non-medically',
-        "self-harm",
-        'presidential national election',
-    ]
-
-    # Iterate over each description in the list
-    for description in descriptions_list:
-        variables_str = f"""
-        {{
-            "where": {{
-                "description": {{
-                    "contains": "{description}"
-                }}
-            }}
-        }}
-        """
-        variables = json.loads(variables_str)
-        graphql_request = {'query': query, 'variables': variables}
-        response = requests.post(graphql_url, json=graphql_request, headers=headers)
-
-        if response.status_code == 200:
-            # Collect each successful responses
-            all_responses.append(response.json().get('data')['measures_A'])
-        else:
-            print("HTTP Error:", response.status_code)
-
-    return all_responses
