+++ python/tests/data/graphql_ahr/golden_data/age_national_current.csv
@@ -7,7 +7,7 @@ United States,00,35-44,,,,,,,,,,,,,18.1
 United States,00,45-54,,,,,,,,,,,,,18.2
 United States,00,55-64,,,,,,,,,,,,,17.0
 United States,00,65+,,8600.0,3.4,20600.0,7100.0,12300.0,14600.0,22600.0,7000.0,8500.0,,,16.9
-United States,00,65-74,,,,,,,,,,,,,15.3
+United States,00,65-74,,,,,,,,,,,1482.0,,15.3
 United States,00,75-84,,,,,,,,,,,,,19.6
 United States,00,85+,,,,,,,,,,,,,22.4
 United States,00,All,,9800.0,8.8,8000.0,3000.0,6200.0,20500.0,10900.0,17300.0,14700.0,2681.0,12.0,14.5
