+++ python/tests/data/graphql_ahr/golden_data/sex_national_current.csv
@@ -1,4 +1,4 @@
 state_name,state_fips,sex,voter_participation_pct_rate,asthma_per_100k,avoided_care_pct_rate,cardiovascular_diseases_per_100k,chronic_kidney_disease_per_100k,copd_per_100k,depression_per_100k,diabetes_per_100k,excessive_drinking_per_100k,frequent_mental_distress_per_100k,preventable_hospitalizations_per_100k,non_medical_drug_use_per_100k,suicide_per_100k
-United States,00,Male,,6800.0,9.1,9400.0,2900.0,5700.0,14300.0,11800.0,20800.0,12000.0,,,23.7
-United States,00,Female,,12400.0,10.6,7300.0,3300.0,7100.0,24400.0,11000.0,13400.0,17200.0,,,5.9
+United States,00,Male,,6800.0,9.1,9400.0,2900.0,5700.0,14300.0,11800.0,20800.0,12000.0,2598.0,,23.7
+United States,00,Female,,12400.0,10.6,7300.0,3300.0,7100.0,24400.0,11000.0,13400.0,17200.0,2756.0,,5.9
 United States,00,All,,9800.0,8.8,8000.0,3000.0,6200.0,20500.0,10900.0,17300.0,14700.0,2681.0,12.0,14.5
