+++ python/tests/datasources/test_graphql_ahr.py
@@ -29,7 +29,7 @@ def _fetch_ahr_data_from_graphql():
 
 
 @mock.patch('ingestion.gcs_to_bq_util.add_df_to_bq', return_value=None)
-@mock.patch('ingestion.gcs_to_bq_util.fetch_ahr_data_from_graphql', side_effect=_fetch_ahr_data_from_graphql)
+@mock.patch('ingestion.graphql_ahr_utils.fetch_ahr_data_from_graphql', side_effect=_fetch_ahr_data_from_graphql)
 def testWriteToBqAgeNational(_mock_fetch: mock.MagicMock, mock_add_df_to_bq: mock.MagicMock):
     datasource = GraphQlAHRData()
     datasource.write_to_bq('dataset', 'gcs_bucket', demographic=AGE, geographic=NATIONAL_LEVEL)
@@ -44,7 +44,7 @@ def testWriteToBqAgeNational(_mock_fetch: mock.MagicMock, mock_add_df_to_bq: moc
 
 
 @mock.patch('ingestion.gcs_to_bq_util.add_df_to_bq', return_value=None)
-@mock.patch('ingestion.gcs_to_bq_util.fetch_ahr_data_from_graphql', side_effect=_fetch_ahr_data_from_graphql)
+@mock.patch('ingestion.graphql_ahr_utils.fetch_ahr_data_from_graphql', side_effect=_fetch_ahr_data_from_graphql)
 def testWriteToBqRaceNational(_mock_fetch: mock.MagicMock, mock_add_df_to_bq: mock.MagicMock):
     datasource = GraphQlAHRData()
     datasource.write_to_bq('dataset', 'gcs_bucket', demographic=RACE_OR_HISPANIC_COL, geographic=STATE_LEVEL)
@@ -59,7 +59,7 @@ def testWriteToBqRaceNational(_mock_fetch: mock.MagicMock, mock_add_df_to_bq: mo
 
 
 @mock.patch('ingestion.gcs_to_bq_util.add_df_to_bq', return_value=None)
-@mock.patch('ingestion.gcs_to_bq_util.fetch_ahr_data_from_graphql', side_effect=_fetch_ahr_data_from_graphql)
+@mock.patch('ingestion.graphql_ahr_utils.fetch_ahr_data_from_graphql', side_effect=_fetch_ahr_data_from_graphql)
 def testWriteToBqSexNational(_mock_fetch: mock.MagicMock, mock_add_df_to_bq: mock.MagicMock):
     datasource = GraphQlAHRData()
     datasource.write_to_bq('dataset', 'gcs_bucket', demographic=SEX, geographic=NATIONAL_LEVEL)
